@@ -4,17 +4,17 @@
     "license": "MIT",
     "type": "library",
     "require": {
-        "php": "^5.5|^7",
+        "php": "^7.1.0",
         "doctrine/dbal": "^2.2",
         "pimple/pimple": "^3.0",
         "silex/api": "^2.0",
-        "doctrine/migrations": "^1.0",
+        "doctrine/migrations": "^1.5",
         "symfony/console": "^2.3 | ^3.0"
     },
     "require-dev": {
         "doctrine/orm": "^2.2",
         "silex/silex": "^2.0",
-        "phpunit/phpunit": "^3.7"
+        "phpunit/phpunit": "^5.7"
     },
     "autoload": {
         "psr-4": { "Kurl\\Silex\\Provider\\": "src/" }
